+++ src/enums.rs
@@ -1,6 +1,8 @@
 // Copy-Paste from https://github.com/pololu/jrk-g2-arduino/blob/master/JrkG2.h
+use ufmt::derive::uDebug;
+
 #[allow(dead_code)]
-#[derive(Debug, Copy, Clone)]
+#[derive(uDebug, Debug, Copy, Clone)]
 pub enum VarOffset {
     Input = 0x00,              // u16
     Target = 0x02,             // u16
@@ -36,7 +38,7 @@ pub enum VarOffset {
 }
 
 #[allow(dead_code)]
-#[derive(Debug, Copy, Clone)]
+#[derive(uDebug, Debug, Copy, Clone)]
 pub enum SettingOffset {
     OptionsByte1 = 0x01,                        // u8
     OptionsByte2 = 0x02,                        // u8
@@ -105,7 +107,7 @@ pub enum SettingOffset {
 }
 
 #[allow(dead_code)]
-#[derive(Debug, Copy, Clone)]
+#[derive(uDebug, Debug, Copy, Clone)]
 pub enum JrkG2Error {
     AwaitingCommand = 0,
     NoPower = 1,
@@ -127,7 +129,7 @@ pub enum JrkG2Error {
 /// I2C interfaces.  These bytes are used by the library and you should not need
 /// to use them.
 #[allow(dead_code)]
-#[derive(Debug, Copy, Clone)]
+#[derive(uDebug, Debug, Copy, Clone)]
 pub enum JrkG2Command {
     SetTarget = 0xC0,
     SetTargetLowResRev = 0xE0,
@@ -151,7 +153,7 @@ pub enum JrkG2Command {
 /// See JrkG2Base::getForceMode(), JrkG2Base::forceDutyCycleTarget(), and
 /// JrkG2Base::forceDutyCycle().
 #[allow(dead_code)]
-#[derive(Debug, Copy, Clone)]
+#[derive(uDebug, Debug, Copy, Clone)]
 pub enum JrkG2ForceMode {
     None = 0,
     DutyCycleTarget = 1,
@@ -163,7 +165,7 @@ pub enum JrkG2ForceMode {
 ///
 /// See JrkG2Base::getDeviceReset().
 #[allow(dead_code)]
-#[derive(Debug, Copy, Clone)]
+#[derive(uDebug, Debug, Copy, Clone)]
 pub enum JrkG2Reset {
     PowerUp = 0,
     Brownout = 1,
@@ -176,7 +178,7 @@ pub enum JrkG2Reset {
 
 /// This enum defines the Jrk G2's control and feedback pins.
 #[allow(dead_code)]
-#[derive(Debug, Copy, Clone)]
+#[derive(uDebug, Debug, Copy, Clone)]
 pub enum JrkG2Pin {
     SCL = 0,
     SDA = 1,
@@ -193,7 +195,7 @@ pub enum JrkG2Pin {
 /// JrkG2Base::getResetIntegral(), JrkG2Base::setCoastWhenOff(), and
 /// JrkG2Base::getCoastWhenOff().
 #[allow(dead_code)]
-#[derive(Debug, Copy, Clone)]
+#[derive(uDebug, Debug, Copy, Clone)]
 pub enum JrkG2OptionsByte3 {
     ResetIntegral = 0,
     CoastWhenOff = 1,