+++ src/jrk.rs
@@ -1,11 +1,13 @@
 use core::fmt;
+use ufmt::{uWrite, uwriteln};
 
 use crate::enums::{JrkG2Command, VarOffset};
 
 pub trait JrkG2<ComError> {
+    const HEADER: &'static str;
+
     fn write(&mut self, data: &[u8]) -> Result<(), ComError>;
     fn read(&mut self, cmd: VarOffset) -> Result<u16, ComError>;
-    fn show_vars_header<W: fmt::Write>(&mut self, f: &mut W);
 
     fn set_target(&mut self, target: u16) -> Result<(), ComError> {
         self.write(&[
@@ -22,7 +24,7 @@ pub trait JrkG2<ComError> {
         Ok(())
     }
     fn show_vars<W: fmt::Write>(&mut self, f: &mut W) -> Result<(), ComError> {
-        self.show_vars_header(f);
+        f.write_str(Self::HEADER).ok();
         self.show_var(f, VarOffset::Input)?;
         self.show_var(f, VarOffset::Target)?;
         self.show_var(f, VarOffset::Feedback)?;
@@ -34,6 +36,27 @@ pub trait JrkG2<ComError> {
         self.show_var(f, VarOffset::ErrorFlagsOccurred)?;
         self.show_var(f, VarOffset::VinVoltage)?;
         self.show_var(f, VarOffset::Current)?;
+        Ok(())
+    }
+
+    // for use with ufmt instead of core::fmt
+    fn ushow_var<W: uWrite>(&mut self, f: &mut W, var: VarOffset) -> Result<(), ComError> {
+        uwriteln!(f, "{:?}: {}", var, self.read(var).ok().unwrap()).ok();
+        Ok(())
+    }
+    fn ushow_vars<W: uWrite>(&mut self, f: &mut W) -> Result<(), ComError> {
+        f.write_str(Self::HEADER).ok();
+        self.ushow_var(f, VarOffset::Input)?;
+        self.ushow_var(f, VarOffset::Target)?;
+        self.ushow_var(f, VarOffset::Feedback)?;
+        self.ushow_var(f, VarOffset::ScaledFeedback)?;
+        self.ushow_var(f, VarOffset::Integral)?;
+        self.ushow_var(f, VarOffset::DutyCycleTarget)?;
+        self.ushow_var(f, VarOffset::PIDPeriodCount)?;
+        self.ushow_var(f, VarOffset::ErrorFlagsHalting)?;
+        self.ushow_var(f, VarOffset::ErrorFlagsOccurred)?;
+        self.ushow_var(f, VarOffset::VinVoltage)?;
+        self.ushow_var(f, VarOffset::Current)?;
         Ok(())
     }
 }